one more pass of clone cleanup
